ENH: text for merges shown in pop up window ENH: start of feature generation: small lines

+++ Tracing/TraceEdit/View3D.h
@@ -118,7 +118,7 @@ protected:
 private:
 	int gapTol;
 	int gapMax;
-	float smallLine;
+	int smallLine;
 	float lineWidth;
 	double SelectColor;
